Can successfully create a budget and send it to Firestore -- and then pull it -- View display isn't implemented yet.

+++ Fidget/Components/CircularProgressView.swift
@@ -8,7 +8,7 @@
 import SwiftUI
 
 struct CircularProgressView: View {
-    var percentage: Float
+    var percentage: Double
     var bgcolor: Color
     var fillColor: Color
     var strokeWidth: Float
